feat: routerを設定とmainをappに分割
