+++ lib/pages/splash.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+class Splash extends StatefulWidget {
+  @override
+  _SplashState createState() => new _SplashState();
+}
+
+class _SplashState extends State<Splash> {
+  @override
+  void initState() {
+    super.initState();
+
+    new Future.delayed(const Duration(seconds: 1))
+        .then((value) => handleTimeout());
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return new Scaffold(
+      body: new Center(
+        // TODO: スプラッシュアニメーション
+        child: const CircularProgressIndicator(),
+      ),
+    );
+  }
+
+  void handleTimeout() {
+    // ログイン画面へ
+    Navigator.of(context).pushReplacementNamed("/login");
+  }
+}
